+++ Assets/Levels/Generation/Noise/INoise3D.cs
@@ -17,7 +17,7 @@ namespace AssemblyCSharp
 		/// </summary>
 		/// <param name="x">The x coordinate.</param>
 		/// <param name="y">The y coordinate.</param>
-		double Noise(double x, double y, double z);
+		float Noise(double x, double y, double z);
 	}
 }
